# os/bluestore: Trim cache on add rather than in loop.