Signed-off-by: Mark Nelson <mnelson@redhat.com>

This PR changes the way that the bluestore cache works to trim onodes and buffers immediately after insertion rather than in a loop inside the mempool thread.  Arguably this still allows the caches to slightly exceed their limits (though by a much smaller amount), however it was easier to implement this way vs freeing the space upfront.  It still results in tighter control of memory, especially for buffer cache.  

Because the cache lock is no longer being taken and held periodically in the mempool thread and instead is being held longer in other threads (fewer locks/unlocks overall), this change has significant effects on bluestore performance and thread/lock contention.  Specifically, performance improved by about 20-25% during a highly concurrent 4K librbd random write workload to an OSD backed by an Intel P3700 NVMe drive:

![FIO+LibRBD 4K RandWrite High-Concurrency (Single-OSD)](https://user-images.githubusercontent.com/1286295/59551600-b5f6b000-8f41-11e9-87f9-dfeef4a75f16.png)

Tail latency (as recorded by fio) also improved dramatically:

| Latency (ms) | Master | wip-bs-cache-evict |
| -- | -- | -- |
| min | 0.305 | 0.308 |
| avg | 8.934625 | 7.094215 |
| 99.9th percentile | 86 | 71 |
| 99.99th percentile | 334 | 113 |
| max | 1142.1 | 571.779 |

Buffer memory is kept on a tighter leash as well.  An interesting side effect is that the priority cache is more aggressively limiting the aggregate cache size to keep total OSD memory usage under control with the new code.  Usually this has a negative effect on performance because a greater percentage of onode reads will have to come from rocksdb block_cache or disk.  In prior testing however it's been observed that higher random write throughput can result in worse cache-memory efficiency which may explain what we are seeing here.  #27705 avoids double caching onodes in the block_cache which may have an even more dramatic effect when combined with this PR.

![OSD Memory Usage during Pre-Fill and 4KB RandWrite Test (Master 2Q Trial 0)](https://user-images.githubusercontent.com/1286295/59551489-d7569c80-8f3f-11e9-9f39-890f6358ddb9.png)
![OSD Memory Usage during Pre-Fill and 4KB RandWrite Test (wip-bs-cache-evict 2Q Trial 0)](https://user-images.githubusercontent.com/1286295/59551488-d7569c80-8f3f-11e9-8bca-c8b1dc3a57ac.png)

To help explain the lower latency and improved performance, gdbpmp was used to look at lock contention behavior before and after the new code was applied.  In master, significant time is spent in the mempool_thread holding the cache lock while performing trim (and for relatively long periods of time):

```
      + 22.90% BlueStore::MempoolThread::_trim_shards
      | + 22.90% BlueStore::Cache::trim
      |   + 22.90% BlueStore::TwoQCache::_trim
      |     + 20.70% BlueStore::Onode::put
      |     | + 20.40% ~Onode
```

However in wip-bs-cache-evict, onode-specific trim is spread across every tp_osd_tp thread and happens more often in shorter bursts:
```
          |     |   |   |   |     + 0.60% BlueStore::TwoQCache::_trim_onodes_to
        | |   |     |   |   | | | + 1.20% BlueStore::TwoQCache::_trim_onodes_to
          |         |   |   | |   + 1.10% BlueStore::TwoQCache::_trim_onodes_to
        | |   | |   | | | | | |   + 0.60% BlueStore::TwoQCache::_trim_onodes_to
          |       | |   |   | | | + 0.70% BlueStore::TwoQCache::_trim_onodes_to
        | |   | | | |   |   | | | + 0.60% BlueStore::TwoQCache::_trim_onodes_to
          |   | |   |   |         + 0.80% BlueStore::TwoQCache::_trim_onodes_to
          |   |     |   |   | | | + 1.80% BlueStore::TwoQCache::_trim_onodes_to
          |         |   | |   |   + 2.10% BlueStore::TwoQCache::_trim_onodes_to
        | |   |     |   |   | |   + 1.20% BlueStore::TwoQCache::_trim_onodes_to
          |   | |   |   |     | | + 1.60% BlueStore::TwoQCache::_trim_onodes_to
          |   |     |   |   | | | + 2.30% BlueStore::TwoQCache::_trim_onodes_to
          |     | | |   |     |   + 1.90% BlueStore::TwoQCache::_trim_onodes_to
        | |   | | | |   |   | | | + 2.60% BlueStore::TwoQCache::_trim_onodes_to
          |     | | |   |     |   + 1.30% BlueStore::TwoQCache::_trim_onodes_to
```

For example, in the new code, onode trim primarily happens when BlueStore::getattr is called:

```
          |     |   |   | + 1.60% BlueStore::getattr
          |     |   |   |   + 1.50% BlueStore::Collection::get_onode
          |     |   |   |   | + 0.90% BlueStore::OnodeSpace::lookup
          |     |   |   |   | | + 0.80% lock_guard
          |     |   |   |   | | | + 0.80% lock
          |     |   |   |   | | |   + 0.80% __gthread_recursive_mutex_lock
          |     |   |   |   | | |     + 0.80% __gthread_mutex_lock
          |     |   |   |   | | |       + 0.80% pthread_mutex_lock
          |     |   |   |   | | |         + 0.80% _L_lock_870
          |     |   |   |   | | |           + 0.80% __lll_lock_wait
          |     |   |   |   | | + 0.10% BlueStore::TwoQCache::_touch_onode
          |     |   |   |   | + 0.60% BlueStore::OnodeSpace::add
          |     |   |   |   |   + 0.60% _trim_onodes
          |     |   |   |   |     + 0.60% BlueStore::TwoQCache::_trim_onodes_to
```

In the original code the mempool thread _trim loop clears both onodes and buffers, but in the new code buffers are trimmed independently of onodes.  Some of this is distributed across the tp_osd_tp threads:

```
          |     |   | |   |   | |       |       + 0.20% BlueStore::TwoQCache::_trim_buffers_to
        | |   |     | | | |   | | | |   | | |   + 0.60% BlueStore::TwoQCache::_trim_buffers_to
          |         | | | |   | |   |   |   |   + 0.30% BlueStore::TwoQCache::_trim_buffers_to
        | |   | |   | | | | | | | | |   |       + 0.30% BlueStore::TwoQCache::_trim_buffers_to
          |       | | | | |   | |   |   |   | | + 0.60% BlueStore::TwoQCache::_trim_buffers_to
        | |   | | | | | | |   | |   |   | |     + 0.30% BlueStore::TwoQCache::_trim_buffers_to
          |   | |   | | | |   | |   |   |       + 0.40% BlueStore::TwoQCache::_trim_buffers_to
          |   |     | |   |   | |   |   | | |   + 0.20% BlueStore::TwoQCache::_trim_buffers_to
          |         | |   |   | |       |       + 0.20% BlueStore::TwoQCache::_trim_buffers_to
        | |   |     | |   |   | |   | | |   |   + 0.10% BlueStore::TwoQCache::_trim_buffers_to
          |   | |   | |   |   | |   |   |       + 0.30% BlueStore::TwoQCache::_trim_buffers_to
          |   |     | | | |   | |   |   |   |   + 0.50% BlueStore::TwoQCache::_trim_buffers_to
          |     | | | | | |   | |   |   |   | | + 0.20% BlueStore::TwoQCache::_trim_buffers_to
        | |   | | | | |   |   | |   |   | | |   + 0.10% BlueStore::TwoQCache::_trim_buffers_to
        | |   | |   | | | | | | |   |   | | |   + 0.30% BlueStore::TwoQCache::_trim_buffers_to
          |     | | | | | |   | |   |   |   |   + 0.40% BlueStore::TwoQCache::_trim_buffers_to
```

While a significant portion of it now happens in bstore_kv_final during _finish_write.  This may be concerning on the surface, but finish_write would have had to wait for the mempool thread to release the cache lock to complete writes anyway as we will see below.

```
        | | | | | + 7.00% BlueStore::TwoQCache::_trim_buffers_to
```

While performance is significantly higher in wip-bs-cache-evict for 4K random writes, more aggregate time is also spent waiting on the cache lock to perform onode lookups. It's possible that this could be due to additional time spent in trim walking over pinned cache items (since onode/buffer trim is called more often now).  It's also possible this is simply due to the fact that we are writing faster and have more opportunities for onode reads to wait on the lock.  It may be worth testing additional cache shards to see if that helps.

Average time the tp_osd_tp threads spent contending on the cache lock during onode lookup:
```
master: 0.89%
wip-bs-cache-evict: 1.90%
```

On the other hand, this PR results in a dramatic reduction in time spent contending on the cache lock in bstore_kv_final during SharedBlob::finish_write.  

master
```
        + 41.40% BlueStore::_txc_state_proc
        | + 31.30% BlueStore::_txc_finish
        | | + 25.20% BlueStore::SharedBlob::finish_write
        | | | + 23.20% lock_guard
        | | | | + 23.20% lock
```
wip-bs-cache-evict:
```
        + 36.30% BlueStore::_txc_state_proc
        | + 25.70% BlueStore::_txc_finish
        | | + 18.90% BlueStore::SharedBlob::finish_write
        | | | + 10.00% lock_guard
        | | | | + 10.00% lock
```

Finish_write no longer needs to wait for (relatively) long periods of time for the mempool thread to perform large onode/buffer cache evictions which drive tail latency up.  Future options to further improve performance may include finer grained locking (independent onode/buffer locks?  More shards?).


- [ ] References tracker ticket
- [ ] Updates documentation if necessary
- [ ] Includes tests for new functionality or reproducer for bug